Moved wko_schema_helper API code
