--- core/src/main/python/wlsdeploy/tool/util/wko/__init__.py
@@ -1,4 +1,0 @@
-"""
-Copyright (c) 2022, Oracle and/or its affiliates.
-Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
-"""
